+++ breeze-data/src/main/kotlin/com/windea/breezeframework/data/dsl/XmlDsl.kt
@@ -11,18 +11,25 @@ import com.windea.breezeframework.data.dsl.XmlConfig.quote
 
 //////////Portal extensions
 
+/**开始生成Xml。*/
 fun Dsl.Companion.xml(builder: Xml.() -> Unit) = Xml.create(builder)
 
+/**开始配置Xml。*/
 fun DslConfig.Companion.xml(builder: XmlConfig.() -> Unit) = XmlConfig.builder()
 
 //////////Config object
 
+/**Xml配置。*/
 object XmlConfig : DslConfig {
+	/**默认根元素名。*/
 	var defaultRootName: String = "root"
 		set(value) = run { field = value.ifBlank { "root" } }
+	/**缩进长度。*/
 	var indentSize = 2
 		set(value) = run { field = value.coerceIn(-2, 8) }
+	/**是否使用双引号。*/
 	var useDoubleQuote: Boolean = true
+	/**是否自关闭标签。*/
 	var autoCloseTag: Boolean = false
 	
 	internal val indent get() = if(indentSize <= -1) "\t" * indentSize else " " * indentSize
@@ -32,17 +39,22 @@ object XmlConfig : DslConfig {
 ///////////////Dsl marker annotations & Dsl element interfaces
 
 @DslMarker
-annotation class XmlDsl
+internal annotation class XmlDsl
 
+/**Xml Dsl元素。*/
 @XmlDsl
 interface XmlDslElement
 
+/**Xml结点。*/
 interface XmlNode : XmlDslElement
 
 ////////////Dsl elements & build functions
 
+/**Xml文件。*/
 class Xml @PublishedApi internal constructor(
+	/**注释列表。*/
 	val comments: MutableList<XmlComment> = mutableListOf(),
+	/**根元素。*/
 	var rootElement: XmlElement = XmlElement(defaultRootName),
 	override var wrapContent: Boolean = true
 ) : Dsl, XmlDslElement, CanWrapContent {
@@ -53,23 +65,32 @@ class Xml @PublishedApi internal constructor(
 	}
 	
 	companion object {
-		inline fun create(builder: Xml.() -> Unit) = Xml().also(builder)
+		@PublishedApi
+		internal inline fun create(builder: Xml.() -> Unit) = Xml().also(builder)
 	}
 	
+	/**添加注释。*/
 	inline fun comment(comment: String) = XmlComment.create(comment).also { comments += it }
 	
+	/**添加元素。*/
 	inline fun element(name: String, vararg attributes: Pair<String, Any?>, builder: XmlElement.() -> Unit) =
 		XmlElement.create(name, *attributes, builder = builder).also { rootElement = it }
 	
+	/**添加注释。*/
 	inline operator fun String.unaryMinus() = comment(this)
 	
+	/**添加元素。*/
 	inline operator fun String.invoke(vararg attributes: Pair<String, Any?>, builder: XmlElement.() -> Unit) =
 		element(this, *attributes, builder = builder)
 }
 
+/**Xml元素。*/
 class XmlElement @PublishedApi internal constructor(
+	/**名字。*/
 	val name: String = "element",
-	val attributes: Map<String, Any?> = mapOf(),
+	/**属性。*/
+	val attributes: MutableMap<String, Any?> = mutableMapOf(),
+	/**子结点列表。*/
 	val nodes: MutableList<XmlNode> = mutableListOf(),
 	override var wrapContent: Boolean = true,
 	override var indentContent: Boolean = true
@@ -90,26 +111,42 @@ class XmlElement @PublishedApi internal constructor(
 	}
 	
 	companion object {
-		inline fun create(name: String, vararg attributes: Pair<String, Any?>, builder: XmlElement.() -> Unit) =
-			XmlElement(name.fixXmlElementName(), attributes.toMap().fixXmlElementAttributes()).also(builder)
+		@PublishedApi
+		internal inline fun create(name: String, vararg attributes: Pair<String, Any?>, builder: XmlElement.() -> Unit) =
+			XmlElement(name.fixXmlElementName(), attributes.toMap().fixXmlElementAttributes().toMutableMap()).also(builder)
 	}
 	
+	/**添加文本。*/
 	inline fun text(text: String) = XmlText.create(text).also { nodes += it }
 	
+	/**添加注释。*/
 	inline fun comment(comment: String) = XmlComment.create(comment).also { nodes += it }
 	
+	/**添加元素。*/
 	inline fun element(name: String, vararg attributes: Pair<String, Any?>, builder: XmlElement.() -> Unit) =
 		XmlElement.create(name, *attributes, builder = builder).also { nodes += it }
 	
+	/**设置属性。*/
+	inline fun attribute(key: String, value: String) = run { this.attributes[key] = value }
+	
+	/**添加文本。*/
 	inline operator fun String.unaryPlus() = text(this)
 	
+	/**添加注释。*/
 	inline operator fun String.unaryMinus() = comment(this)
 	
+	/**添加元素。*/
 	inline operator fun String.invoke(vararg attributes: Pair<String, Any?>, builder: XmlElement.() -> Unit) =
 		element(this, *attributes, builder = builder)
+	
+	/**设置属性。*/
+	inline operator fun String.invoke(value: String) = attribute(this, value)
 }
 
+
+/**Xml文本。*/
 class XmlText @PublishedApi internal constructor(
+	/**文本。*/
 	var text: String = ""
 ) : XmlNode {
 	override fun toString(): String {
@@ -117,11 +154,14 @@ class XmlText @PublishedApi internal constructor(
 	}
 	
 	companion object {
-		inline fun create(text: String) = XmlText(text.fixXmlText())
+		@PublishedApi
+		internal inline fun create(text: String) = XmlText(text.fixXmlText())
 	}
 }
 
+/**Xml注释。*/
 class XmlComment @PublishedApi internal constructor(
+	/**注释。*/
 	var comment: String = "",
 	override var wrapContent: Boolean = false,
 	override var indentContent: Boolean = false
@@ -138,7 +178,8 @@ class XmlComment @PublishedApi internal constructor(
 	}
 	
 	companion object {
-		inline fun create(comment: String) = XmlComment(comment.fixXmlText())
+		@PublishedApi
+		internal inline fun create(comment: String) = XmlComment(comment.fixXmlText())
 	}
 }