+++ CHANGELOG.md
@@ -44,6 +44,7 @@
 * [X] breeze-data 代理实现`DataSerializer`的相关类。
 * [X] breeze-reflect 添加`checkClassForName`方法。
 * [X] breeze-core 添加核心扩展方法：`once`。
+* [X] breeze-core 添加函数式扩展方法的示例：`curried`、`parial`
 
 ### 1.0.2
 
@@ -52,8 +53,6 @@
 * [X] breeze-core 移除Range的中缀扩展，因为可能返回`ClosedRange<T>`或`Pair`。
 * [X] breeze-core 将更多简单的扩展方法标为内联。（为了完全消除性能影响）
 * [X] breeze-data 完成一个更好的Xml Dsl示例。
-* [ ] 尝试提供线程安全（双重校验锁）的单例模式接口。
-    * （kotlin的object实现的是懒汉式的单例，但提供对于方法的`@Synchronized`注解）
 * [ ] DELAY breeze-core Uri构建和编码的扩展。
 * [ ] DELAY breeze-core 利用反射的类型转换的扩展。
 * [ ] DELAY breeze-text en 单数与复数的转化。